♻️ [Refactor] entities 파일 구조 수정
<body>
- models/common: base entity types
- models/dtos: dto types
- models/tables: table types

<footer>
- 관련: #3 #8
