+++ src/models/tables/board.entity.ts
@@ -1,6 +1,6 @@
 import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
-import { BoardStatus } from '../types/enums/board-status.enum';
-import { CommonEntity } from './common/common.entity';
+import { BoardStatus } from '../../types/enums/board-status.enum';
+import { CommonEntity } from '../common/common.entity';
 
 @Entity()
 export class BoardEntity extends CommonEntity {
