+++ src/repositories/board.repository.ts
@@ -1,6 +1,6 @@
 import { Repository } from 'typeorm';
 import { CustomRepository } from '../configs/custom-typeorm.decorator';
-import { BoardEntity } from '../entities/board.entity';
+import { BoardEntity } from '../models/tables/board.entity';
 
 @CustomRepository(BoardEntity)
 export class BoardRespository extends Repository<BoardEntity> {}
