+++ src/services/boards.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { BoardEntity } from '../entities/board.entity';
-import { CreateBoardDto } from '../entities/dtos/create-board.dto';
+import { BoardEntity } from '../models/tables/board.entity';
+import { CreateBoardDto } from '../models/dtos/create-board.dto';
 import { BoardRespository } from '../repositories/board.repository';
 import { BoardStatus } from '../types/enums/board-status.enum';
